+++ frontend/components/feedback-task/footer/Pagination.component.vue
@@ -1,57 +1,34 @@
 <template>
   <div class="pagination">
-    <div class="number-of-records-by-page-area"></div>
     <div class="pagination__buttons">
       <BaseButton
         class="pagination__button"
         ref="prevButton"
-        @click="onPaginate(onClickPrev)"
-        :disabled="currentPage === 1"
+        @click="onClickPrev"
+        :disabled="isFirstPage"
       >
         <svgicon name="chevron-left" width="8" height="8" />
         {{ prevButtonMessage }}
       </BaseButton>
 
-      <div class="pagination__page-number-area" v-if="showPageNumber">
-        <BaseButton
-          v-for="page in totalPages"
-          :key="page"
-          class="pagination__button"
-          @click="onClickNumber(page)"
-          :disabled="isCurrentPage(page)"
-        >
-          {{ page }}
-        </BaseButton>
-      </div>
-
       <BaseButton
         class="pagination__button"
         ref="nextButton"
-        @click="onPaginate(onClickNext)"
-        :disabled="currentPage >= totalPages"
+        @click="onClickNext"
+        :disabled="false"
       >
         {{ nextButtonMessage }}
         <svgicon name="chevron-right" width="8" height="8" />
       </BaseButton>
     </div>
-    <!-- NOTE: HIDDEN FOR MVP
-    <div class="total-records-area">
-      <span v-text="totalOfRecordMessage" />
-    </div> -->
   </div>
 </template>
 
 <script>
-import { Notification } from "@/models/Notifications";
-
 export default {
   name: "PaginationComponent",
   props: {
-    totalItems: {
-      type: Number,
-      required: true,
-    },
-    numberOfItemsByPage: {
+    currentPage: {
       type: Number,
       default: () => 1,
     },
@@ -63,59 +40,21 @@ export default {
       type: String,
       default: () => "Prev",
     },
-    showPageNumber: {
-      type: Boolean,
-      default: false,
-    },
-    notificationParams: {
-      type: Object | null,
-    },
-    conditionToShowNotificationComponentOnPagination: {
-      type: Function | null,
-    },
-  },
-  data() {
-    return {
-      localCurrentPage: 1,
-    };
   },
   computed: {
-    currentPage: {
-      get() {
-        return this.localCurrentPage;
-      },
-      set(newCurrentPage) {
-        this.localCurrentPage = +newCurrentPage;
-      },
+    isFirstPage() {
+      return this.currentPage === 1;
     },
-    totalPages() {
-      return Math.ceil(this.totalItems / this.numberOfItemsByPage);
-    },
-    totalOfRecordMessage() {
-      return `${this.currentPage} of ${this.totalItems} records`;
-    },
-    isPageAvailable() {
-      return this.pageFromRoute <= this.totalPages;
-    },
+    // TODO: Move to the parent
     pageFromRoute() {
       return parseFloat(this.$route.query?._page) || 1;
     },
   },
-  watch: {
-    localCurrentPage: {
-      immediate: true,
-      handler(newCurrentPage) {
-        this.emitCurrentPage();
-        this.updateUrlParams(newCurrentPage);
-      },
-    },
-  },
   mounted() {
-    this.currentPage = this.isPageAvailable ? this.pageFromRoute : 1;
     document.addEventListener("keydown", this.onPressKeyboardShortCut);
-    this.onBusEventCurrentPage();
   },
   destroyed() {
+    // TODO: Move this to the parent
     document.removeEventListener("keydown", this.onPressKeyboardShortCut);
   },
   methods: {
@@ -135,72 +74,11 @@ export default {
         // Do nothing => the code is not registered as shortcut
       }
     },
-    onBusEventCurrentPage() {
-      this.$root.$on("current-page", (currentPage) => {
-        this.currentPage = currentPage;
-      });
-    },
-    onPaginate(eventToFire) {
-      if (this.isNotificationComponentForThisPage()) {
-        const { message, buttonMessage, typeOfToast } = this.notificationParams;
-        this.showNotificationBeforePaginate({
-          eventToFire,
-          message,
-          buttonMessage,
-          typeOfToast,
-        });
-      } else {
-        eventToFire();
-      }
-    },
-    isNotificationComponentForThisPage() {
-      if (
-        this.notificationParams &&
-        this.conditionToShowNotificationComponentOnPagination
-      ) {
-        const showNotification =
-          this.conditionToShowNotificationComponentOnPagination();
-
-        return showNotification;
-      }
-      return false;
-    },
-    showNotificationBeforePaginate({
-      eventToFire,
-      message,
-      buttonMessage,
-      typeOfToast,
-    }) {
-      Notification.dispatch("notify", {
-        message: message ?? "",
-        numberOfChars: 500,
-        type: typeOfToast ?? "warning",
-        buttonText: buttonMessage ?? "",
-        async onClick() {
-          eventToFire();
-        },
-      });
-    },
     onClickPrev() {
-      this.currentPage > 1 && this.currentPage--;
+      this.$emit("on-click-prev");
     },
     onClickNext() {
-      this.currentPage < this.totalPages && this.currentPage++;
-    },
-    onClickNumber(pageValue) {
-      this.currentPage = pageValue;
-    },
-    emitCurrentPage() {
-      this.$emit("on-paginate", this.currentPage);
-    },
-    isCurrentPage(page) {
-      return this.currentPage === page;
-    },
-    updateUrlParams(currentPage) {
-      this.$router.push({
-        path: this.$route.path,
-        query: { ...this.$route.query, _page: currentPage },
-      });
+      this.$emit("on-click-next");
     },
   },
 };