+++ frontend/components/page-contents/feedback-task-content/RecordFeedbackTaskAndQuestionnaire.content.vue
@@ -1,9 +1,5 @@
 <template>
-  <div
-    v-if="!$fetchState.pending && !$fetchState.error"
-    :key="recordOffset"
-    class="wrapper"
-  >
+  <div v-if="!$fetchState.pending && !$fetchState.error" class="wrapper">
     <template v-if="recordId">
       <RecordFeedbackTaskComponent
         v-if="fieldsWithRecordFieldText"
@@ -11,6 +7,7 @@
         :fields="fieldsWithRecordFieldText"
       />
       <QuestionsFormComponent
+        :key="questionFormKey"
         class="question-form"
         :class="statusClass"
         v-if="questionsWithRecordAnswers && questionsWithRecordAnswers.length"
@@ -18,11 +15,15 @@
         :recordId="recordId"
         :recordStatus="record.record_status"
         :initialInputs="questionsWithRecordAnswers"
+        @on-submit-responses="onActionInResponses('SUBMIT')"
+        @on-discard-responses="onActionInResponses('DISCARD')"
+        @on-question-form-touched="onQuestionFormTouched"
       />
     </template>
+
     <div v-else class="wrapper--empty">
       <p
-        v-if="!totalRecords"
+        v-if="!hasRecords"
         class="wrapper__text --heading3"
         v-text="noRecordsMessage"
       />
@@ -32,6 +33,8 @@
 </template>
 
 <script>
+import { isNil } from "lodash";
+import { Notification } from "@/models/Notifications";
 import {
   getQuestionsByDatasetId,
   getComponentTypeOfQuestionByDatasetIdAndQuestionName,
@@ -42,14 +45,14 @@ import {
   RECORD_STATUS,
   RESPONSE_STATUS_FOR_API,
   upsertRecords,
+  deleteAllRecords,
   getRecordWithFieldsAndResponsesByUserId,
   isRecordWithRecordIndexByDatasetIdExists,
+  isAnyRecordByDatasetId,
 } from "@/models/feedback-task-model/record/record.queries";
-import {
-  updateTotalRecordsByDatasetId,
-  getTotalRecordByDatasetId,
-} from "@/models/feedback-task-model/feedback-dataset/feedbackDataset.queries";
+
 import { COMPONENT_TYPE } from "@/components/feedback-task/feedbackTask.properties";
+import { LABEL_PROPERTIES } from "../../feedback-task/feedbackTask.properties";
 
 const TYPE_OF_FEEDBACK = Object.freeze({
   ERROR_FETCHING_RECORDS: "ERROR_FETCHING_RECORDS",
@@ -59,10 +62,6 @@ export default {
   name: "RecordFeedbackTaskAndQuestionnaireComponent",
   props: {
     datasetId: {
-      type: String,
-      required: true,
-    },
-    recordStatusToFilterWith: {
       type: String,
       required: true,
     },
@@ -78,14 +77,35 @@ export default {
         return { orderFieldsBy: "order", ascendent: true };
       },
     },
-    recordOffset: {
-      type: Number,
-      required: true,
+  },
+  watch: {
+    async currentPage(newValue) {
+      await this.$router.push({
+        path: this.$route.path,
+        query: {
+          ...this.$route.query,
+          _page: newValue,
+          _status: this.recordStatusToFilterWith,
+        },
+      });
+    },
+    async recordStatusToFilterWith(newValue) {
+      await this.$router.push({
+        path: this.$route.path,
+        query: {
+          ...this.$route.query,
+          _status: newValue,
+          _page: this.currentPage,
+        },
+      });
     },
   },
   data() {
     return {
-      rerenderQuestionnaire: 1,
+      reRenderQuestionForm: 1,
+      questionFormTouched: false,
+      recordStatusToFilterWith: null,
+      currentPage: null,
     };
   },
   computed: {
@@ -111,15 +131,15 @@ export default {
       }
       return paramForUrl;
     },
-    totalRecords() {
-      return getTotalRecordByDatasetId(this.datasetId);
-    },
     record() {
       return getRecordWithFieldsAndResponsesByUserId(
         this.datasetId,
         this.userId,
-        this.recordOffset
+        this.currentPage - 1
       );
+    },
+    hasRecords() {
+      return isAnyRecordByDatasetId(this.datasetId);
     },
     recordId() {
       return this.record?.id;
@@ -180,50 +200,165 @@ export default {
     statusClass() {
       return `--${this.record.record_status.toLowerCase()}`;
     },
-  },
-  async fetch() {
-    await this.initRecordsInDatabase(this.recordOffset);
-  },
-  watch: {
-    recordOffset(newRecordOffset) {
-      const isRecordWithRecordOffsetNotExists =
-        !isRecordWithRecordIndexByDatasetIdExists(
-          this.datasetId,
-          newRecordOffset
-        );
-      if (isRecordWithRecordOffsetNotExists) {
-        this.initRecordsInDatabase(newRecordOffset);
-      }
+    questionFormKey() {
+      return `${this.currentPage}-${this.reRenderQuestionForm}`;
     },
+    statusFilterFromQuery() {
+      return this.$route.query?._status ?? RECORD_STATUS.PENDING.toLowerCase();
+    },
+    pageFromQuery() {
+      const { _page } = this.$route.query;
+      return isNil(_page) ? 1 : +_page;
+    },
+  },
+
+  async fetch() {
+    await this.initRecordsInDatabase(this.currentPage - 1);
+
+    const offset = this.currentPage - 1;
+    const isRecordExistForCurrentPage =
+      isRecordWithRecordIndexByDatasetIdExists(this.datasetId, offset);
+
+    if (!isRecordExistForCurrentPage) {
+      await this.initRecordsInDatabase(0);
+      this.currentPage = 1;
+    }
+  },
+  async created() {
+    this.noMoreDataMessage = "There is no more data";
+
+    this.recordStatusToFilterWith = this.statusFilterFromQuery;
+    this.currentPage = this.pageFromQuery;
+  },
+  mounted() {
+    this.$root.$on("go-to-next-page", () => {
+      this.setCurrentPage(this.currentPage + 1);
+    });
+    this.$root.$on("go-to-prev-page", () => {
+      this.setCurrentPage(this.currentPage - 1);
+    });
+    this.$root.$on("status-filter-changed", this.onStatusFilterChanged);
   },
   methods: {
-    async initRecordsInDatabase(recordOffset) {
-      // FETCH records from recordOffset + 5 next records
-      const { items: records, total: totalRecords } = await this.getRecords(
+    async applyStatusFilter(status) {
+      this.recordStatusToFilterWith = status;
+      this.currentPage = 1;
+
+      await deleteAllRecords();
+      await this.$fetch();
+
+      this.reRenderQuestionForm++;
+      this.questionFormTouched = false;
+    },
+    emitResetStatusFilter() {
+      this.$root.$emit("reset-status-filter");
+    },
+    async onStatusFilterChanged(newStatus) {
+      if (this.recordStatusToFilterWith === newStatus) {
+        return;
+      }
+
+      const localApplyStatusFilter = this.applyStatusFilter;
+      const localEmitResetStatusFilter = this.emitResetStatusFilter;
+
+      if (this.questionFormTouched) {
+        Notification.dispatch("notify", {
+          message: "Your changes will be lost if you move to another view",
+          numberOfChars: 500,
+          type: "warning",
+          buttonText: LABEL_PROPERTIES.CONTINUE,
+          async onClick() {
+            await localApplyStatusFilter(newStatus);
+          },
+          onClose() {
+            localEmitResetStatusFilter();
+          },
+        });
+      } else {
+        await this.applyStatusFilter(newStatus);
+      }
+    },
+    onQuestionFormTouched(isTouched) {
+      this.questionFormTouched = isTouched;
+    },
+    async setCurrentPage(newPage) {
+      const offset = newPage - 1;
+      let isNextRecordExist = isRecordWithRecordIndexByDatasetIdExists(
         this.datasetId,
-        recordOffset,
-        5
+        offset
       );
 
-      // FORMAT records for orm
-      const formattedRecords = this.factoryRecordsForOrm(records);
+      // if isNextRecordExist => move to the next one
+      // else fetch new set of records and move to next one
+      if (!isNextRecordExist) {
+        await this.initRecordsInDatabase(offset);
+        isNextRecordExist = isRecordWithRecordIndexByDatasetIdExists(
+          this.datasetId,
+          offset
+        );
+      }
 
-      // UPSERT total records && records in ORM
-      updateTotalRecordsByDatasetId(this.datasetId, totalRecords);
-      upsertRecords(formattedRecords);
+      if (isNextRecordExist) {
+        this.currentPage = newPage;
+      } else if (this.currentPage < newPage) {
+        // notify there is no more records
+        Notification.dispatch("notify", {
+          message: this.noMoreDataMessage,
+          numberOfChars: this.noMoreDataMessage.length,
+          type: "info",
+        });
+      }
     },
-    async getRecords(datasetId, recordOffset, numberOfRecordsToFetch = 5) {
+    async onActionInResponses(typeOfAction) {
+      switch (typeOfAction) {
+        case "SUBMIT":
+        case "DISCARD":
+          this.setCurrentPage(this.currentPage + 1);
+          break;
+        default:
+        // do nothing
+      }
+    },
+    async initRecordsInDatabase(
+      offset,
+      status = this.recordStatusFilterValueForGetRecords
+    ) {
+      // FETCH records from offset, status + 10 next records
+      const { items: records } = await this.getRecords(
+        this.datasetId,
+        offset,
+        status
+      );
+      // FORMAT records for orm
+      const formattedRecords = this.factoryRecordsForOrm(records, offset);
+
+      // UPSERT records in ORM
+      await upsertRecords(formattedRecords);
+    },
+    async getRecords(
+      datasetId,
+      offset,
+      responseStatus,
+      numberOfRecordsToFetch = 10
+    ) {
       try {
-        const url = `/v1/me/datasets/${datasetId}/records?include=responses&offset=${recordOffset}&limit=${numberOfRecordsToFetch}&response_status=${this.recordStatusFilterValueForGetRecords}`;
-        const { data } = await this.$axios.get(url);
+        const url = `/v1/me/datasets/${datasetId}/records`;
+        const params = {
+          include: "responses",
+          offset,
+          limit: numberOfRecordsToFetch,
+          response_status: responseStatus,
+        };
+        const { data } = await this.$axios.get(url, { params });
         return data;
       } catch (err) {
+        console.warn(err);
         throw {
           response: TYPE_OF_FEEDBACK.ERROR_FETCHING_RECORDS,
         };
       }
     },
-    factoryRecordsForOrm(records) {
+    factoryRecordsForOrm(records, offset) {
       return records.map(
         (
           { id: recordId, responses: recordResponses, fields: recordFields },
@@ -240,7 +375,7 @@ export default {
 
           return {
             id: recordId,
-            record_index: index + this.recordOffset,
+            record_index: index + offset,
             dataset_id: this.datasetId,
             record_status: recordStatus ?? RECORD_STATUS.PENDING,
             record_fields: formattedRecordFields,
@@ -344,6 +479,11 @@ export default {
 
       return { formattedRecordResponsesForOrm, recordStatus };
     },
+  },
+  beforeDestroy() {
+    this.$root.$off("go-to-next-page");
+    this.$root.$off("go-to-prev-page");
+    this.$root.$off("status-filter-changed");
   },
 };
 </script>