+++ frontend/models/feedback-task-model/feedback-dataset/feedbackDataset.queries.js
@@ -5,13 +5,6 @@ const upsertFeedbackDataset = (feedbackDataset) => {
   FeedbackDatasetModel.insertOrUpdate({ data: feedbackDataset });
 };
 
-const updateTotalRecordsByDatasetId = (datasetId, totalRecords) => {
-  FeedbackDatasetModel.update({
-    where: datasetId,
-    data: { total_records: totalRecords },
-  });
-};
-
 // GET
 const getAllFeedbackDatasets = () => {
   return FeedbackDatasetModel.all();
@@ -24,9 +17,6 @@ const getFeedbackDatasetWorkspaceNameById = (datasetId) => {
     FeedbackDatasetModel.query().whereId(datasetId).first()?.workspace_name ||
     null
   );
-};
-const getTotalRecordByDatasetId = (datasetId) => {
-  return FeedbackDatasetModel.query().whereId(datasetId).first()?.total_records;
 };
 const getDatasetTaskByDatasetId = (datasetId) => {
   return FeedbackDatasetModel.query().whereId(datasetId).first()?.task;
@@ -46,13 +36,11 @@ const deleteDatasetById = (datasetId) => {
 };
 export {
   upsertFeedbackDataset,
-  updateTotalRecordsByDatasetId,
   getAllFeedbackDatasets,
   getFeedbackDatasetNameById,
   getDatasetGuidelinesByDatasetId,
   getDatasetTaskByDatasetId,
   getFeedbackDatasetWorkspaceNameById,
-  getTotalRecordByDatasetId,
   isDatasetByIdExists,
   deleteDatasetById,
 };