+++ src/argilla/__init__.py
@@ -19,6 +19,7 @@ as well as in the `_import_structure` dictionary.
 """
 
 import sys as _sys
+import warnings
 from typing import TYPE_CHECKING as _TYPE_CHECKING
 
 from argilla.logging import configure_logging as _configure_logging
@@ -33,6 +34,15 @@ try:
     _install_rich()
 except ModuleNotFoundError:
     pass
+
+# TODO: Remove this warning once https://github.com/argilla-io/argilla/issues/2902 is tackled
+if _sys.version_info < (3, 8):
+    warnings.warn(
+        message="Python 3.7 is coming to its end-of-life and will be no longer supported in the upcoming release of Argilla. "
+        "To ensure compatibility and uninterrupted service, we kindly request that you migrate to Argilla with"
+        " Python 3.8 or higher.",
+        category=DeprecationWarning,
+    )
 
 __version__ = _version.version
 
@@ -62,7 +72,6 @@ if _TYPE_CHECKING:
         RatingQuestion,
         TextField,
         TextQuestion,
-        create_feedback_dataset,
     )
     from argilla.client.models import (
         Text2TextRecord,
@@ -114,7 +123,6 @@ _import_structure = {
     ],
     "client.feedback": [
         "FeedbackDataset",
-        "create_feedback_dataset",
         "TextField",
         "TextQuestion",
         "RatingQuestion",