+++ src/argilla/client/datasets.py
@@ -1499,6 +1499,9 @@ Dataset = Union[DatasetForTextClassification, DatasetForTokenClassification, Dat
 def read_datasets(dataset: "datasets.Dataset", task: Union[str, TaskType], **kwargs) -> Dataset:
     """Reads a datasets Dataset and returns a argilla Dataset
 
+    Columns not supported by the :mod:`Record <argilla.client.models>` instance corresponding
+    with the task are ignored.
+
     Args:
         dataset: Dataset to be read in.
         task: Task for the dataset, one of: ["TextClassification", "TokenClassification", "Text2Text"].
@@ -1555,6 +1558,9 @@ def read_datasets(dataset: "datasets.Dataset", task: Union[str, TaskType], **kwa
 
 def read_pandas(dataframe: pd.DataFrame, task: Union[str, TaskType]) -> Dataset:
     """Reads a pandas DataFrame and returns a argilla Dataset
+
+    Columns not supported by the :mod:`Record <argilla.client.models>` instance corresponding
+    with the task are ignored.
 
     Args:
         dataframe: Dataframe to be read in.