+++ src/argilla/client/sdk/v1/datasets/models.py
@@ -42,7 +42,7 @@ class FeedbackValueModel(BaseModel):
 class FeedbackResponseModel(BaseModel):
     id: UUID
     values: Dict[str, FeedbackValueModel]
-    status: Literal["submitted", "missing", "discarded"]
+    status: Literal["submitted", "discarded"]  # API also contains "missing", but it's just a filter-status
     user_id: UUID
     inserted_at: datetime
     updated_at: datetime
@@ -59,10 +59,6 @@ class FeedbackItemModel(BaseModel):
 
 class FeedbackRecordsModel(BaseModel):
     items: List[FeedbackItemModel]
-    total: Optional[int] = None
-
-    class Config:
-        fields = {"total": {"exclude": True}}
 
 
 class FeedbackFieldModel(BaseModel):